+++ lib/httpi/adapter/net_http.rb
@@ -127,6 +127,10 @@ module HTTPI
         request_client = request_class.new @request.url.request_uri, @request.headers
         request_client.basic_auth *@request.auth.credentials if @request.auth.basic?
 
+        if @request.auth.digest?
+          raise NotSupportedError, "Net::HTTP does not support HTTP digest authentication"
+        end
+
         request_client
       end